% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDSeroMonitor.R
\name{FMDSeroMonitor}
\alias{FMDSeroMonitor}
\title{###############################################################################}
\usage{
FMDSeroMonitor(Prevac, Postvac, Census_Data)
}
\arguments{
\item{Prevac}{A n X 3 data frame (n: number of states in the sample) obtained from sero-}

\item{Postvac}{A n X 3 data frame (n: number of states in the sample) obtained from sero-}

\item{Census_Data}{Animal census N X 1 data frame (state wise bovine population), where rows}
}
\value{
it will give the state estimate, national estimate, vaccination impact
}
\description{
###############################################################################
}
\examples{
# Example usage (ensure 'prevac', 'postvac', and 'CensusData' are available in your environment)

}
