% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FMDSeroMonitor.R
\name{SeroMonitor}
\alias{SeroMonitor}
\title{###############################################################################}
\usage{
SeroMonitor(Prevac, Postvac, Census_Data)
}
\arguments{
\item{Prevac}{A n X 3 data frame (n: number of states in the sample) obtained from sero-}

\item{Postvac}{A n X 3 data frame (n: number of states in the sample) obtained from sero-}

\item{Census_Data}{Animal census N X 1 data frame (state wise bovine population), where rows}
}
\description{
###############################################################################
}
